pyproject.toml
[project]
name = "hands-off-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "openai-agents>=0.0.17",
    "tavily-python>=0.7.6",
]
